+++ scripts/esm.js
@@ -27,7 +27,7 @@ const ENTRIES = {
     'tools.mjs': '../src/tools.js',
     'projects.mjs': '../src/projects.js',
     'vip.mjs': '../src/vip.js',
-    'chrome.mjs': '../src/chrome.js',
+    'frame.mjs': '../src/frame.js',
     'testing.mjs': '../src/testing/index.js',
 };